python-container-dev/requirements.txt
Flask
numpy
moviepy
Pillow
scikit-video
ffmpeg-python
yt-dlp
awscli
requests
pymkv
boto3==1.24.89
opencv-python==4.8.1.78
imageio==2.22.1
av==10.0.0